+++ src/utils.ts
@@ -1,9 +1,11 @@
 import { treeGraph } from './tree-graph.js';
 import exitWithError from './exit-with-error.js';
 import countSubcommands from './utils/count-commands.js';
+import { navigateSubcommands } from './utils/navigate-subcommands.js';
 
 export {
     exitWithError,
     treeGraph,
     countSubcommands,
+    navigateSubcommands,
 };